Move TestAssessmentBase into it's own module
